backend/requirements.txt
appdirs==1.4.4
# apsw==3.36.0.post1
# apturl==0.5.2
asgiref==3.7.2
attrs==23.1.0
autobahn>=22.4.2,<23
Automat==22.10.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
# Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
channels==4.0.0
chardet==4.0.0
# chrome-gnome-shell==0.0.0
click==8.0.3
cloudinary==1.33.0
colorama==0.4.4
# command-not-found==0.3
constantly==15.1.0
cryptography==41.0.3
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
# cupshelpers==1.0
cycler==0.11.0
daphne==4.0.0
# dbus-python==1.2.18
decorator==4.4.2
# defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.7.0
# distro-info===1.1build1
Django==4.2.4
django-allauth==0.54.0
django-chartjs==2.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.1.0
# duplicity==0.8.21
fasteners==0.14.1
feedparser==6.0.8
flake8==4.0.1
flatbuffers==23.5.26
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
# gpg==1.16.0
# gyp==0.1
html2text==2020.1.16
# html5-parser==0.4.10
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
# idna==3.3
ifaddr==0.1.7
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.6.4
incremental==22.10.0
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
jwcrypto==1.5.0
keyring==23.5.0
kiwisolver==1.3.2
# language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
# louis==3.20.0
lxml==4.8.0
# lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
# mpmath==0.0.0
msgpack==1.0.3
netifaces==0.11.0
# numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
proglog==0.1.10
prompt-toolkit==3.0.28
protobuf==3.12.4
psycopg==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
# py7zr==0.11.3+dfsg
pyasn1==0.5.0
pyasn1-modules==0.3.0
# pycairo==1.20.1
pyChart.JS==0.3.0
# pychm==0.8.6
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
# pycups==2.0.1
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.11.2
# PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
# python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
# python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python3-openid==3.2.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
# PyYAML==5.4.1
regex==2021.11.10
reportlab==3.6.8
repoze.lru==0.7
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
Routes==2.5.1
scipy==1.8.0
SecretStorage==3.3.1
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.1
sqlparse==0.4.4
sympy==1.9
# systemd-python==234
texttable==1.6.4
tqdm==4.65.0
traitlets==5.1.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
# ubuntu-advantage-tools==8001
# ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
# ufw==0.36.1
# unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
# usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
wrapt==1.15.0
xdg==5
# xkit==0.0.0
zeroconf==0.38.3
zipp==1.0.0
zope.interface==6.0
